+++ test/obs_websocket_test.dart
@@ -5,7 +5,6 @@ import 'package:test/test.dart';
 
 void main() {
   test('GetStreamStatus', () async {
-    expect(true, true);
     final response = json.decode(
         '{"d":{"requestId":"80cf6055-23be-4241-982f-b75ef6194047","requestStatus":{"code":100,"result":true},"requestType":"GetStreamStatus","responseData":{"outputActive":false,"outputBytes":0,"outputCongestion":0.0,"outputDuration":0,"outputReconnecting":false,"outputSkippedFrames":0,"outputTimecode":"00:00:00.000","outputTotalFrames":0}},"op":7}');